Fix arithmetic error in skipRows()
Fixes an arithmetic error and changes ints to longs to mimic C code
(even though the argument passed in is a long cast to an int)

+++ src/main/java/com/teradata/tpcds/random/RandomNumberStream.java
@@ -20,7 +20,7 @@ public interface RandomNumberStream
 
     double nextRandomDouble();
 
-    void skipRows(int numberOfRows);
+    void skipRows(long numberOfRows);
 
     void resetSeed();
